+++ commerce/delivery/src/main/java/ru/yandex/practicum/service/DeliveryService.java
@@ -3,8 +3,6 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.delivery.dto.DeliveryDto;
 import ru.yandex.practicum.order.dto.OrderDto;
 
-
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -17,5 +15,5 @@ public interface DeliveryService {
 
     void failDelivery(UUID orderId);
 
-    BigDecimal calculateDeliveryCost(OrderDto orderDto);
+    Double calculateDeliveryCost(OrderDto orderDto);
 }
