+++ commerce/order/src/main/java/ru/yandex/practicum/service/OrderServiceImpl.java
@@ -9,7 +9,6 @@ import ru.yandex.practicum.model.Order;
 import ru.yandex.practicum.order.enums.OrderStatus;
 import ru.yandex.practicum.repository.OrderRepository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -78,7 +77,7 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional
-    public Order setTotalPrice(UUID orderId, BigDecimal totalCost) {
+    public Order setTotalPrice(UUID orderId, Double totalCost) {
         Order order = findOrderById(orderId);
         order.setTotalPrice(totalCost);
 
@@ -87,7 +86,7 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     @Transactional
-    public Order setDeliveryPrice(UUID orderId, BigDecimal deliveryCost) {
+    public Order setDeliveryPrice(UUID orderId, Double deliveryCost) {
         Order order = findOrderById(orderId);
         order.setDeliveryPrice(deliveryCost);
