+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -1,22 +1,20 @@
 package ru.yandex.practicum.service;
 
 
-
 import ru.yandex.practicum.order.dto.OrderDto;
 import ru.yandex.practicum.payment.dto.PaymentDto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto order);
 
-    BigDecimal getTotalCost(OrderDto order);
+    Double getTotalCost(OrderDto order);
 
     void paymentSuccess(UUID orderId);
 
-    BigDecimal getProductCost(OrderDto order);
+    Double getProductCost(OrderDto order);
 
     void paymentFailed(UUID orderId);
 }
